Telegram bot now can receive a fill
Still need to see if this is the best way of handling commands/menus
