+++ pkg/presenters/telegram/n26MonthTracking.go
@@ -16,53 +16,66 @@ type DocumentConfig struct {
 
 const EXIT_COMMAND = "N26TrackingClose"
 
-func setOneTimeKeyBoardMap(items []string, rowsCant int) tgbotapi.ReplyKeyboardMarkup {
-	matrix := [][]tgbotapi.KeyboardButton{}
-	row := []tgbotapi.KeyboardButton{}
-	for _, category := range items {
-		newButton := tgbotapi.NewKeyboardButton(category)
-		row = append(row, newButton)
-		if len(row) == rowsCant || len(items)-len(matrix)*rowsCant == len(row) {
-			matrix = append(matrix, row)
-			row = []tgbotapi.KeyboardButton{}
-		}
+func displayCategories(tbot *tgbotapi.BotAPI, chatID int64, q *querying.Service) error {
+	cg := q.CategoryQuerier
+	resp, err := cg.Query()
+	if err != nil {
+		return err
 	}
-	return tgbotapi.NewOneTimeReplyKeyboard(matrix...)
+	categories := []string{}
+	for _, name := range resp.Categories {
+		categories = append(categories, name)
+	}
+	msg := tgbotapi.NewMessage(chatID, "Pick a category")
+	msg.ReplyMarkup = setOneTimeKeyBoardMap(categories, 4)
+	tbot.Send(msg)
+	return nil
+}
+
+func categorizeAndCreate(tbot *tgbotapi.BotAPI, chatID int64, newUpdate tgbotapi.Update, file *tgbotapi.File, t *tracking.Service) error {
+	return nil
 }
 
 func n26MonthTracking(tbot *tgbotapi.BotAPI, update *tgbotapi.Update, updates *tgbotapi.UpdatesChannel, t *tracking.Service, q *querying.Service) {
 	tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, "Starting N26 Report 📄"))
 	tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, "Please send me an N26 csv file export"))
+	// Iterate over the new updates
 	for newUpdate := range *updates {
+		chatID := update.Message.Chat.ID
 		if newUpdate.Message.Document != nil {
-			fileID := newUpdate.Message.Document.FileID
-			file, err := tbot.GetFile(tgbotapi.FileConfig{FileID: fileID})
+			file, err := readSentFile(tbot, chatID, *update)
 			if err != nil {
-				tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, err.Error()))
+				tbot.Send(tgbotapi.NewMessage(chatID, err.Error()))
 				break
 			}
-			fmt.Printf("Received file '%s' with ID %s", file.FilePath, fileID)
-			url := "https://api.telegram.org/file/bot" + tbot.Token + "/" + file.FilePath
-			tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, fmt.Sprintf("You can download the file here: %s", url)))
-			// TODO: Move this into a new function for better readability.
-			cg := q.CategoryQuerier
-			resp, err := cg.Query()
+			err = displayCategories(tbot, chatID, q)
 			if err != nil {
-				tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, err.Error()))
+				tbot.Send(tgbotapi.NewMessage(chatID, err.Error()))
+				break
 			}
-			category_keys := []string{}
-			for _, name := range resp.Categories {
-				category_keys = append(category_keys, name)
+			err = categorizeAndCreate(tbot, chatID, newUpdate, file, t)
+			if err != nil {
+				tbot.Send(tgbotapi.NewMessage(chatID, err.Error()))
+				break
 			}
-			msg := tgbotapi.NewMessage(update.Message.Chat.ID, "Pick a category")
-			msg.ReplyMarkup = setOneTimeKeyBoardMap(category_keys, 4)
-			tbot.Send(msg)
-			// END TODO
-			break
 		} else {
 			tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, "Closing N26 Report Importer 📄"))
 			tbot.Send(tgbotapi.NewMessage(update.Message.Chat.ID, "Please send me an N26 csv file export. N26 Monlty tracker stopped."))
 			break
 		}
 	}
+}
+
+func readSentFile(tbot *tgbotapi.BotAPI, chatID int64, newUpdate tgbotapi.Update) (*tgbotapi.File, error) {
+	msg := tgbotapi.NewMessage(chatID, "")
+	fileID := newUpdate.Message.Document.FileID
+	file, err := tbot.GetFile(tgbotapi.FileConfig{FileID: fileID})
+	if err != nil {
+		return nil, err
+	}
+	fmt.Printf("Received file '%s' with ID %s", file.FilePath, fileID)
+	url := "https://api.telegram.org/file/bot" + tbot.Token + "/" + file.FilePath
+	msg.Text = fmt.Sprintf("You can download the file here: \n %s\n %s", file.Link(tbot.Token), url)
+	tbot.Send(msg)
+	return &file, nil
 }
